rename talks_spec file
